chore(documentation-v7): remove demo stories
