+++ packages/documentation-v7/src/stories/components/card/card.stories.tsx
@@ -408,3 +408,27 @@ export const CardGroup: Story = {
   `;
   },
 }
+
+export const BackgroundImage: Story = {
+  ...singleCardStory,
+  parameters: {
+    controls: {
+      include: [],
+    },
+  },
+  args: {
+    showImage: false,
+    customBody: `<img class="card-img" src="https://picsum.photos/id/20/300/200" alt="Card image" />
+  <div class="card-img-overlay">
+    <div class="card-body">
+      <h5 class="card-title">Titulum</h5>
+
+      <p class="card-text">Contentus momentus vero siteos et accusam iretea et justo.</p>
+
+      <button class="btn btn-primary btn-animated">
+        <span>Butonon teksto</span>
+      </button>
+    </div>
+  </div>`
+  }
+}
