+++ packages/documentation-v7/src/stories/utilities/heading/heading.stories.ts
@@ -1,8 +1,12 @@
 import type { Meta, StoryObj } from '@storybook/web-components';
 import { html } from 'lit';
+import { BADGE } from '../../../../.storybook/constants';
 
 const meta: Meta = {
-  title: 'hidden/demos/utilities/Heading',
+  title: 'Utilities/Heading',
+  parameters: {
+    badges: [BADGE.NEEDS_REVISION],
+  },
 };
 
 export default meta;
