+++ packages/documentation-v7/src/stories/utilities/list/list.docs.mdx
@@ -1,11 +1,7 @@
 import { Meta, Canvas } from '@storybook/blocks';
-import { BADGE } from '../../../../.storybook/constants';
-import * as ListDemos from './list.demo.stories';
+import * as ListStories from './list.stories';
 
-<Meta
-  title="Utilities/List"
-  parameters={{ badges: [BADGE.TODO] }}
-/>
+<Meta of={ListStories} />
 
 # Lists
 
@@ -19,21 +15,21 @@ Utility classes allow you to style each of these lists in different ways.
 An unordered list groups related items in no specific order or sequence.
 It is created using a `<ul>` tag and each list item is preceded by a bullet (small circle).
 
-<Canvas of={ListDemos.UnorderedList} sourceState='shown'/>
+<Canvas of={ListStories.UnorderedList} sourceState="shown" />
 
 ### Unstyled
 
 Remove the default list-style and left margin on list items (immediate children only).
 This only applies to immediate child list items, meaning you will need to add the class for any nested lists as well.
 
-<Canvas of={ListDemos.UnstyledList} sourceState='shown'/>
+<Canvas of={ListStories.UnstyledList} sourceState="shown" />
 
 ### Inline
 
 Remove a list’s bullets and apply some light margin with a combination of two classes,
 `.list-inline` and `.list-inline-item`.
 
-<Canvas of={ListDemos.InlineList} sourceState="shown"/>
+<Canvas of={ListStories.InlineList} sourceState="shown"/>
 
 ## Ordered List
 
@@ -42,18 +38,18 @@ It is created using a `<ol>` tag and list items are preceded by ascending number
 
 <strong>All of the above utility classes also apply to ordered lists.</strong>
 
-<Canvas of={ListDemos.OrderedList} sourceState='shown' />
+<Canvas of={ListStories.OrderedList} sourceState="shown" />
 
 ## Description List
 
 A description list is a list of terms, with a description of each term.
 It is created using a `<dl>` tag, terms are defined using `<dt>` tags, and descriptions are defined using `<dd>` tags.
 
-<Canvas of={ListDemos.DescriptionList} sourceState='shown' />
+<Canvas of={ListStories.DescriptionList} sourceState="shown" />
 
 ### Alignment
 
 Align terms and descriptions horizontally by using our grid system’s predefined classes.
 For longer terms, you can optionally add a `.text-truncate` class to truncate the text with an ellipsis.
 
-<Canvas of={ListDemos.DescriptionListUsingGrid} sourceState='shown' />
+<Canvas of={ListStories.DescriptionListUsingGrid} sourceState="shown" />
